Refactor database schema and entities for Document and User management
- Updated `Document` entity to link with `DocumentType` and `DocumentVersion` instead of storing file paths directly.
- Replaced direct entity associations with explicit join tables (`DocumentTag`, `DocumentClassification`).
- Renamed `ApplicationSettings` entity and repository to `ApplicationSetting`.
- Upgraded primary key data types in multiple entities (e.g., from `Integer` to `Long`) for consistency.
- Updated JPA mappings with explicit foreign keys, unique constraints, and fetch types across all entities (`User`, `Role`, `Tag`, `Classification`, `Bookmark`, `Group`).
- Refactored `BulkUploadService` and `BulkUploadTemplateGenerator` to support the new metadata fields (`Document Type`, `Version Number`).
- Modified `StoredProcedureRepository` to align with the updated database stored procedures (`sp_UploadDocument`, `sp_DeleteDocument`, `sp_UpdateDocument`).
- Updated corresponding DTOs (`DocumentDto`, `DocumentInfoDTO`, `DocumentMetadata`, `UserDto`, etc.) to reflect the new schema changes.
- Removed `pageNumber` logic from bookmarks.

+++ src/main/java/codesAndStandards/springboot/registrationlogin/controller/PandocHtmlController.java
@@ -39,7 +39,7 @@ public class PandocHtmlController {
 
     @PreAuthorize("hasAnyAuthority('Manager', 'Admin', 'Viewer')")
     @GetMapping("/documents/pandoc-preview/{id}")
-    public ResponseEntity<?> pandocPreview(@PathVariable Integer id) {
+    public ResponseEntity<?> pandocPreview(@PathVariable Long id) {
         File tempDir   = null;
         File inputFile = null;
         try {
