+++ src/main/java/codesAndStandards/springboot/registrationlogin/repository/AccessControlLogicRepository.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 @Repository
 public interface AccessControlLogicRepository extends JpaRepository<AccessControlLogic, Long> {
@@ -28,7 +27,7 @@ public interface AccessControlLogicRepository extends JpaRepository<AccessContro
     @Query("SELECT acl FROM AccessControlLogic acl " +
             "JOIN FETCH acl.group " +
             "WHERE acl.document.id = :documentId")
-    List<AccessControlLogic> findByDocumentId(@Param("documentId") int documentId);
+    List<AccessControlLogic> findByDocumentId(@Param("documentId") Long documentId);
 
     /**
      * Check if document is in a group
@@ -36,14 +35,14 @@ public interface AccessControlLogicRepository extends JpaRepository<AccessContro
     @Query("SELECT CASE WHEN COUNT(acl) > 0 THEN true ELSE false END " +
             "FROM AccessControlLogic acl " +
             "WHERE acl.document.id = :documentId AND acl.group.id = :groupId")
-    boolean existsByDocumentIdAndGroupId(@Param("documentId") int documentId, @Param("groupId") Long groupId);
+    boolean existsByDocumentIdAndGroupId(@Param("documentId") Long documentId, @Param("groupId") Long groupId);
 
     /**
      * Find specific document-group association
      */
     @Query("SELECT acl FROM AccessControlLogic acl " +
             "WHERE acl.document.id = :documentId AND acl.group.id = :groupId")
-    Optional<AccessControlLogic> findByDocumentIdAndGroupId(@Param("documentId") int documentId, @Param("groupId") Long groupId);
+    Optional<AccessControlLogic> findByDocumentIdAndGroupId(@Param("documentId") Long documentId, @Param("groupId") Long groupId);
 
     /**
      * Delete all documents from a group
@@ -53,19 +52,18 @@ public interface AccessControlLogicRepository extends JpaRepository<AccessContro
     void deleteByGroupId(@Param("groupId") Long groupId);
 
     /**
-     * ✅ NEW: Delete all group associations for a specific document
-     * Used when updating a document's group assignments
+     * Delete all group associations for a specific document
      */
     @Modifying
     @Query("DELETE FROM AccessControlLogic acl WHERE acl.document.id = :documentId")
-    void deleteByDocumentId(@Param("documentId") int documentId);
+    void deleteByDocumentId(@Param("documentId") Long documentId);
 
     /**
      * Delete specific document from a group
      */
     @Modifying
     @Query("DELETE FROM AccessControlLogic acl WHERE acl.document.id = :documentId AND acl.group.id = :groupId")
-    void deleteByDocumentIdAndGroupId(@Param("documentId") int documentId, @Param("groupId") Long groupId);
+    void deleteByDocumentIdAndGroupId(@Param("documentId") Long documentId, @Param("groupId") Long groupId);
 
     /**
      * Count documents in a group
@@ -86,10 +84,7 @@ public interface AccessControlLogicRepository extends JpaRepository<AccessContro
             "FROM AccessControlLogic acl " +
             "JOIN GroupUser gu ON acl.group.id = gu.group.id " +
             "WHERE gu.user.id = :userId AND acl.document.id = :documentId")
-    boolean hasUserAccessToDocument(@Param("userId") Long userId, @Param("documentId") int documentId);
-
-//    @Query("SELECT acl FROM AccessControlLogic acl WHERE acl.document.id = :documentId")
-//    List<AccessControlLogic> findByDocumentId(@Param("documentId") int documentId);
+    boolean hasUserAccessToDocument(@Param("userId") Long userId, @Param("documentId") Long documentId);
 
     /**
      * Get all document IDs accessible by a user
